requirements.txt
requests
pandas
numpy
ta
schedule
python-dotenv
